feat: Add contestant status to participants
Introduces a new `isContestant` boolean field to participant data. This allows distinguishing participants who are also contestants in an event.

The UI has been updated to include this field in the participant form and display it in the participant detail view. The import process also now supports mapping a 'Contestant' column to this new status.